--- src/controllers/error404.controller.js
@@ -1,7 +1,0 @@
-import views from "../views/error404.html";
-
-export default () => {
-  const divElement = document.createElement("div");
-  divElement.innerHTML = views;
-  return divElement;
-};
