--- src/controllers/home.controller.js
@@ -1,7 +1,0 @@
-import views from "../views/home.html";
-
-export default () => {
-  const divElement = document.createElement("div");
-  divElement.innerHTML = views;
-  return divElement;
-};
