+++ src/index.js
@@ -1,0 +1,7 @@
+import "bootstrap/dist/css/bootstrap.min.css";
+import "./main.scss";
+import router from "./routes/route";
+router(window.location.hash);
+window.addEventListener("hashchange", () => {
+  router(window.location.hash);
+});
