+++ src/modules/about.js
@@ -1,0 +1,7 @@
+import views from "../views/about.html";
+
+export default () => {
+  const divElement = document.createElement("div");
+  divElement.innerHTML = views;
+  return divElement;
+};
